+++ doc.go
@@ -36,7 +36,7 @@ Source code and other details for the project are available at GitHub:
 
 Current Version
 
-0.0.7
+0.0.8
 
 Installation
 
@@ -220,9 +220,10 @@ Example Code:
 	func main() {
 		// Let's add a custom level,
 		//
-		// It should be starting from level with uint 5, 0 is the first.
-		// because we have 5 built'n levels:
+		// It should be starting from level index 6,
+		// because we have 6 built'n levels  (0 is the start index):
 		// disable,
+		// fatal,
 		// error,
 		// warn,
 		// info
@@ -230,7 +231,7 @@ Example Code:
 
 		// First we create our level to a golog.Level
 		// in order to be used in the Log functions.
-		var SuccessLevel golog.Level = 5
+		var SuccessLevel golog.Level = 6
 		// Register our level, just three fields.
 		golog.Levels[SuccessLevel] = &golog.LevelMetadata{
 			Name:    "success",
@@ -259,6 +260,7 @@ passing its string representation to the `SetLevel` function.
 Example Code:
 
 	golog.SetLevel("disable")
+	golog.SetLevel("fatal")
 	golog.SetLevel("error")
 	golog.SetLevel("warn")
 	golog.SetLevel("info")
@@ -392,4 +394,4 @@ Examples:
 package golog // import "github.com/kataras/golog"
 
 // Version is the version string representation of the "golog" package.
-const Version = "0.0.7"
+const Version = "0.0.8"